# Parallel Rubocop: added "check:rubocop" and "check:rubocop:auto_correct" tasks