--- a/package/rubygem-yast-rake.changes
+++ b/package/rubygem-yast-rake.changes
@@ -1,4 +1,12 @@
 -------------------------------------------------------------------
+Fri Jun 22 14:05:30 UTC 2018 - lslezak@suse.cz
+
+- Added "check:rubocop" and "check:rubocop:auto_correct" tasks
+  which run Rubocop in parallel (much faster that the traditional
+  sequential way)
+- 0.2.26
+
+-------------------------------------------------------------------
 Fri Jun 15 13:07:57 UTC 2018 - jreidinger@suse.com
 
 - switch sle_latest to SLE15 SP1 and switch SLE15 GA to