- The new tasks run Rubocop in parallel which is much faster that the traditional sequential way.
- 0.2.26

### Speed Up in yast2-storage-ng

- Running Rubocop locally: from 44s to 12s (~3.6x faster on a quad core CPU)
- In Travis: from 69s to 47s (~1.5x faster - maximum possible, Travis has about 1.5x CPU build power)