@lslezak 2018-06-22
@jreidinger please merge and release the new gem, thank you!